requirements.txt
pytest==7.4.0
requests==2.31.0
allure-pytest==2.13.2